Added unit and turret pricing

+++ backend/AbstractUnitFactory/Units/Unit.cs
@@ -8,7 +8,7 @@ namespace tower_battle.AbstractUnitFactory.Units
         public float Speed { get; set; }
         public Vector2 Position { get; set; }
         public abstract Vector2 Scale { get; set; }
-        public int Cost { get; set; }
+        public double Cost { get; set; }
         public double InitialHealth { get; set; }
 
         public double Health { get; set; }
